+++ libs/react/src/__mocks__/productsMock.ts
@@ -44,6 +44,11 @@ export const getProductsMock = (): ProductsQuery[] => {
       metadata: null,
       seasonCode: '123456789',
       upcCode: '123456789',
+      brandProductId: '123456789',
+      supplierProductId: '123456789',
+      description: 'Description 1',
+      productCategory: 'Product Category 1',
+      productSubcategory: 'Product Subcategory 1',
     },
     {
       id: '2',
@@ -73,6 +78,11 @@ export const getProductsMock = (): ProductsQuery[] => {
       metadata: null,
       seasonCode: '123456789',
       upcCode: '123456789',
+      brandProductId: '123456789',
+      supplierProductId: '123456789',
+      description: 'pneumonoultramicroscopicsilicovolcanoconiosis Odio quis eum non ullam voluptatem quas. Aut hic laboriosam nesciunt nisi omnis voluptas expedita. Quos iusto totam doloribus molestiae eum dolores sint ipsam. Maxime fuga neque qui ipsam tempore. Quisquam quia exercitationem voluptatum et earum enim beatae exercitationem quas. Nihil voluptatem maiores vel hic necessitatibus cumque.',
+      productCategory: null,
+      productSubcategory: 'Product Subcategory 1',
     },
     {
       id: '3',
@@ -102,6 +112,11 @@ export const getProductsMock = (): ProductsQuery[] => {
       metadata: null,
       seasonCode: '123456789',
       upcCode: '123456789',
+      brandProductId: '123456789',
+      supplierProductId: '123456789',
+      description: 'Description 1',
+      productCategory: 'Product Category 1',
+      productSubcategory: null,
     },
   ]
 }