+++ libs/react/src/components/organisms/productsDataGrid/productsDataGrid.tsx
@@ -18,12 +18,17 @@ const getColumnRows = (products: ProductsQuery[]) => {
     return {
       id: product.id,
       name: product.name,
+      category: product.productCategory ?? '',
+      subcategory: product.productSubcategory ?? '',
+      description: product.description ?? '',
       sustainabilityAttributes: uniq(product.attributeAssurances.map(attribute => attribute.sustainabilityAttribute.name)),
       tier1Supplier: product.organizationFacility?.name ?? '',
-      materials: product.productMaterials.map(material => material.material?.name).filter((material): material is string => material !== null),
-      upcCode: product.upcCode ?? '',
-      styleCode: product.styleCode ?? '',
       seasonCode: product.seasonCode ?? '',
+      styleCode: product.styleCode ?? '',
+      upcCode: product.upcCode ?? '',
+      brandProductId: product.brandProductId ?? '',
+      supplierProductId: product.supplierProductId ?? '',
+      materials: product.productMaterials.map(material => material.material?.name).filter((material): material is string => material !== null),
     }
   })
 }
@@ -45,12 +50,54 @@ export const _ProductsDataGrid = () => {
     flex: 1,
   }
 
+  const renderName = (params: any) => {
+    const name = get(params, 'row.name', '')
+    const category = get(params, 'row.category', '')
+    const subcategory = get(params, 'row.subcategory', '')
+    const text = [category, subcategory]
+			.filter((i: string) => (i !== ''))
+			.join(' | ');
+
+    return (
+      <div className={'flex flex-col w-full h-full justify-center gap-[8px]'}>
+        <div className={'w-full h-auto items-center text-body font-bold truncate'}>
+          <span>{name}</span>
+        </div>
+        {
+          text &&
+          <div className={'w-full h-auto items-center text-body text-tc-disabled truncate'}>
+            <span>{text}</span>
+          </div>
+        }
+      </div>
+    )
+  }
+
+  const renderDescription = (params: any) => {
+    const description = get(params, 'row.description', '')
+    return (
+      <div className={'w-full h-full text-body items-center flex text-wrap break-words'}>
+        <span className={'line-clamp-3'}>{description}</span>
+      </div>
+    )
+  }
+
   const columns = [
     {
       ...defaultColumnProperties,
       field: 'name',
       headerName: 'Name',
-      width: 200
+      minWidth: 200,
+      flex: 1,
+      renderCell: renderName,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'description',
+      headerName: 'Description',
+      minWidth: 200,
+      flex: 1,
+      renderCell: renderDescription,
     },
     {
       ...defaultColumnProperties,
@@ -68,7 +115,43 @@ export const _ProductsDataGrid = () => {
       ...defaultColumnProperties,
       field: 'tier1Supplier',
       headerName: 'Tier 1 Supplier',
-      width: 200,
+      minWidth: 200,
+      flex: 1,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'seasonCode',
+      headerName: 'Season',
+      minWidth: 200,
+      flex: 1,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'styleCode',
+      headerName: 'Style',
+      minWidth: 200,
+      flex: 1,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'upcCode',
+      headerName: 'UPC',
+      minWidth: 200,
+      flex: 1,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'brandProductId',
+      headerName: 'Brand Product ID',
+      minWidth: 200,
+      flex: 1,
+    },
+    {
+      ...defaultColumnProperties,
+      field: 'supplierProductId',
+      headerName: 'Supplier Product ID',
+      minWidth: 200,
+      flex: 1,
     },
     {
       ...defaultColumnProperties,
@@ -80,24 +163,6 @@ export const _ProductsDataGrid = () => {
           <DataGridCellHoverPopover params={params} />
         )
       },
-    },
-    {
-      ...defaultColumnProperties,
-      field: 'upcCode',
-      headerName: 'UPC',
-      width: 200,
-    },
-    {
-      ...defaultColumnProperties,
-      field: 'styleCode',
-      headerName: 'Style',
-      width: 200,
-    },
-    {
-      ...defaultColumnProperties,
-      field: 'seasonCode',
-      headerName: 'Season',
-      width: 200,
     },
   ]
 
@@ -157,7 +222,7 @@ export const _ProductsDataGrid = () => {
         columns={columns}
         rows={rows}
         slots={{ toolbar: getToolbar }}
-        rowHeight={55}
+        rowHeight={114}
       />
     </div>
   )