+++ libs/react/src/lib/graphql/queries.ts
@@ -218,6 +218,9 @@ export const GET_ALL_PRODUCTS= gql`
     products(filter: $filter){
       id
       name
+      productCategory
+      productSubcategory
+      description
       productMaterials {
         id
         material {
@@ -253,6 +256,8 @@ export const GET_ALL_PRODUCTS= gql`
       metadata
       seasonCode
       upcCode
+      brandProductId
+      supplierProductId
     }
   }
 `;